requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
supabase==2.4.6
fal-client==0.4.1
razorpay==1.4.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-multipart==0.0.9
setuptools
